% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_recruitment_time}
\alias{add_recruitment_time}
\title{Add recruitment time to a survival dataset}
\usage{
add_recruitment_time(
  data,
  rec_method,
  rec_period = NULL,
  rec_power = NULL,
  rec_rate = NULL,
  rec_duration = NULL
)
}
\arguments{
\item{data}{A survival dataframe with columns: time, status, group}

\item{rec_method}{Recruitment method: "power" or "PWC" (piecewise constant)}

\item{rec_period}{Period length for power model}

\item{rec_power}{Power parameter for power model}

\item{rec_rate}{Comma-separated string of rates for PWC model}

\item{rec_duration}{Comma-separated string of durations for PWC model}
}
\value{
A dataframe with added columns: rec_time and pseudo_time
}
\description{
Add recruitment time to a survival dataset
}
