% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{assurance_GSD_shiny_app}
\alias{assurance_GSD_shiny_app}
\title{Shiny app which simulates group sequential trials with DTE, using elicited distributions}
\usage{
assurance_GSD_shiny_app()
}
\description{
You should run the function with no arguments.
}
