% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_BPP_hist}
\alias{calc_BPP_hist}
\title{Calculates the Bayesian Predictive Probability for a trial with a Delayed Treatment Effect}
\usage{
calc_BPP_hist(
  n_c,
  n_t,
  control_dist = "Exponential",
  t1 = NULL,
  t2 = NULL,
  t1_Beta_a = NULL,
  t1_Beta_b = NULL,
  diff_Beta_a = NULL,
  diff_Beta_b = NULL,
  delay_time_SHELF,
  delay_time_dist = "hist",
  post_delay_HR_SHELF,
  post_delay_HR_dist = "hist",
  P_S = 1,
  P_DTE = 0,
  cens_events = NULL,
  IF = NULL,
  rec_method,
  rec_period = NULL,
  rec_power = NULL,
  rec_rate = NULL,
  rec_duration = NULL,
  alpha_spending = NULL,
  beta_spending = NULL,
  IF_list = NULL,
  k = 2,
  N = 50,
  M = 50
)
}
\arguments{
\item{n_c}{Number of patients in the control group}

\item{n_t}{Number of patients in the treatment group}

\item{control_dist}{Distribution of control group, must be one of "Exponential" (default) or "weibull"}

\item{t1}{Time 1}

\item{t2}{Time 2}

\item{t1_Beta_a}{Hyperparameter a for the Beta distribution for the survival probability at time 1}

\item{t1_Beta_b}{Hyperparameter a for the Beta distribution for the survival probability at time 1}

\item{diff_Beta_a}{Hyperparameter a for the Beta distribution for the difference in survival probabilities (t2-t1)}

\item{diff_Beta_b}{Hyperparameter b for the Beta distribution for the difference in survival probabilities (t2-t1)}

\item{delay_time_SHELF}{A SHELF object, beliefs about the delay time}

\item{delay_time_dist}{Distribution of the delay time, "hist" is default. See SHELF help for more details}

\item{post_delay_HR_SHELF}{A SHELF object, beliefs about the post-delay hazard ratio}

\item{post_delay_HR_dist}{Distribution of the post-delay hazard ratio, "hist" is default. See SHELF help for more details}

\item{P_S}{Probability of the survival curves separating}

\item{P_DTE}{Probability of the survival curves being subject to a DTE, given they separate}

\item{cens_events}{Number of events at which you wish to perform the censoring (must be less than n_c + n_t)}

\item{IF}{Information fraction at which BPP is calculated}

\item{rec_method}{Recruitment method, must be one of "power" or "PWC" (piecewise constant)}

\item{rec_period}{Parameter used to model recruitment according to power model}

\item{rec_power}{Parameter used to model recruitment according to power model}

\item{rec_rate}{Parameter used to model recruitment according to piecewise constant model}

\item{rec_duration}{Parameter used to model recruitment according to piecewise constant model}

\item{alpha_spending}{Cumulative alpha spending}

\item{beta_spending}{Cumulative beta spending}

\item{IF_list}{Vector of information fractions - must be in the format: `c("0.5, 1")'}

\item{k}{Number of stages}

\item{N}{Number of trials simulated (default is 50)}

\item{M}{Number of samples from the posterior (default is 50)}
}
\value{
Numeric vector of length \code{N} containing the simulated Bayesian Predictive Probabilities.
}
\description{
Calculates the Bayesian Predictive Probability for a trial with a Delayed Treatment Effect
}
\examples{
set.seed(123)
# Minimal example for CRAN: small n_sim for fast execution
res <- calc_BPP_hist(n_c = 300,
                     n_t = 300,
                     control_dist = "Exponential",
                     t1 = 12,
                     t1_Beta_a = 20,
                     t1_Beta_b = 32,
                     delay_time_SHELF = SHELF::fitdist(c(5.5, 6, 6.5), probs = c(0.25, 0.5, 0.75), lower = 0, upper = 12),
                     delay_time_dist = "gamma",
                     post_delay_HR_SHELF = SHELF::fitdist(c(0.5, 0.6, 0.7), probs = c(0.25, 0.5, 0.75), lower = 0, upper = 1),
                     post_delay_HR_dist = "gamma",
                     P_S = 1,
                     P_DTE = 0,
                     cens_events = 300,
                     rec_method = "power",
                     rec_period = 12,
                     rec_power = 1,
                     alpha_spending = c(0.01, 0.025),
                     beta_spending = c(0.05, 0.1),
                     IF_list = c("0.5, 1"),
                     k = 2,
                     M = 5,
                     N = 5)
str(res)

}
