% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calibrate_BPP_thresholds}
\alias{calibrate_BPP_thresholds}
\title{Function to calculate the 'optimal' BPP threshold value}
\usage{
calibrate_BPP_thresholds(
  n_c,
  n_t,
  control_model,
  effect_model,
  recruitment_model,
  IA_model,
  analysis_model,
  data_generating_model,
  n_sims = 100
)
}
\arguments{
\item{n_c}{Number of control patients}

\item{n_t}{Number of treatment patients}

\item{control_model}{A named list specifying the control arm survival distribution:
\itemize{
\item \code{dist}: Distribution type ("Exponential" or "Weibull")
\item \code{parameter_mode}: Either "Fixed" or "Distribution"
\item \code{fixed_type}: If "Fixed", specify as "Parameters" or "Landmark"
\item \code{lambda}, \code{gamma}: Scale and shape parameters
\item \code{t1}, \code{t2}: Landmark times
\item \code{surv_t1}, \code{surv_t2}: Survival probabilities at landmarks
\item \code{t1_Beta_a}, \code{t1_Beta_b}, \code{diff_Beta_a}, \code{diff_Beta_b}: Beta prior parameters
}}

\item{effect_model}{A named list specifying beliefs about the treatment effect:
\itemize{
\item \code{delay_SHELF}, \code{HR_SHELF}: SHELF objects encoding beliefs
\item \code{delay_dist}, \code{HR_dist}: Distribution types ("hist" by default)
\item \code{P_S}: Probability that survival curves separate
\item \code{P_DTE}: Probability of delayed separation, conditional on separation
}}

\item{recruitment_model}{A named list specifying the recruitment process:
\itemize{
\item \code{method}: "power" or "PWC"
\item \code{period}, \code{power}: Parameters for power model
\item \code{rate}, \code{duration}: Comma-separated strings for PWC model
}}

\item{IA_model}{A named list specifying the censoring mechanism for the future data:
\itemize{
\item \code{events}: Number of events which is 100\% information fraction
\item \code{IF}: The information fraction at which to censor and calculate BPP
}}

\item{analysis_model}{A named list specifying the final analysis and decision rule:
\itemize{
\item \code{method}: e.g. \code{"LRT"}, \code{"WLRT"}, or \code{"MW"}.
\item \code{alpha}: one-sided type I error level.
\item \code{alternative_hypothesis}: direction of the alternative (e.g. \code{"one.sided"}).
\item \code{rho}, \code{gamma}, \code{t_star}, \code{s_star}: additional parameters for WLRT or MW (if applicable).
}}

\item{data_generating_model}{A named list specifying the data-generating mechanisms
\itemize{
\item \code{null_model}: \code{lambda_c}
\item \code{fixed_delay}: \code{lambda_c}, \code{delay_time} and \code{post_delay_HR}
\item \code{no_delay}: \code{lambda_c} and \code{HR}
}}

\item{n_sims}{Number of data sets to simulate (default is 100).}
}
\value{
A vector of length \code{n_sims} corresponding to the value of BPP for each simulated trial
}
\description{
Function to calculate the 'optimal' BPP threshold value
}
