% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensFunc.R
\name{CensFunc}
\alias{CensFunc}
\title{Function to censor a data set}
\usage{
CensFunc(dataCombined, censTime)
}
\arguments{
\item{dataCombined}{a data frame}

\item{censTime}{Time at whcih you wish to perform the censoring}
}
\value{
A dataframe
}
\description{
Function to censor a data set
}
