\name{INTEREST}
\docType{data}
\alias{INTEREST}
\title{INTEREST clinical trial data set}
\description{
  Scraped Kaplan-Meier data from the clinical trial: Gefitinib versus docetaxel in previously treated non-small-cell lung cancer (INTEREST): a randomised phase III trial
}
\usage{INTEREST}
\format{A data set containing 710 observations.}
\references{
  Kim ES, Hirsh V, Mok T, et al. Gefitinib versus docetaxel in previously treated non-small-cell lung cancer (INTEREST): a randomised phase III trial. Lancet. 2008;372(9652):1809-1818. doi:10.1016/S0140-6736(08)61758-4
}
\keyword{datasets}
