% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCMC_sample}
\alias{MCMC_sample}
\title{MCMC_sample}
\format{
A data frame with 100000 rows and 1 variables:
\describe{
\item{x}{Sample from the MAP prior}

}
}
\source{
A MCMC sample for the control group for the example given in https://onlinelibrary.wiley.com/doi/full/10.1002/sim.10136.
Three historical data seta are used to generate a Meta-Analytic-Predictive Prior distribution
}
\usage{
MCMC_sample
}
\description{
An MCMC sample for the example given in Salsbury et al (2024)
}
\keyword{datasets}
