\name{REVEL}
\docType{data}
\alias{REVEL}
\title{REVEL clinical trial data set}
\description{
  Scraped Kaplan-Meier data from the clinical trial: Ramucirumab plus docetaxel versus placebo plus docetaxel for second-line treatment of stage IV non-small-cell lung cancer after disease progression on platinum-based therapy (REVEL): a multicentre, double-blind, randomised phase 3 trial
}
\usage{REVEL}
\format{A data set containing 625 observations.}
\references{
 Garon EB, Ciuleanu TE, Arrieta O, et al. Ramucirumab plus docetaxel versus placebo plus docetaxel for second-line treat-
ment of stage IV non-small-cell lung cancer after disease progression on platinum-based therapy (REVEL): a multicentre,
double-blind, randomised phase 3 trial. Lancet. 2014;384(9944):665-673. doi:10.1016/S0140-6736(14)60845-X
}
\keyword{datasets}
