% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDTEDataSet.R
\name{SimDTEDataSet}
\alias{SimDTEDataSet}
\title{Function to simulate a data set, in the context of delayed treatment effects}
\usage{
SimDTEDataSet(
  n_C,
  n_E,
  lambda_C,
  lambda_E,
  gamma_C,
  gamma_E,
  delayT,
  rec_method,
  rec_period = NULL,
  rec_power = NULL,
  rec_rate = NULL,
  rec_duration = NULL
)
}
\arguments{
\item{n_C}{Sample size in control group}

\item{n_E}{Sample size in experimental treatment group}

\item{lambda_C}{Weibull control parameter}

\item{lambda_E}{Weibull experimental treatment parameter}

\item{gamma_C}{Weibull control parameter}

\item{gamma_E}{Weibull experimental treatment parameter}

\item{delayT}{Length of delay}

\item{rec_method}{Recruitment method}

\item{rec_period}{Recruitment period (for the power method)}

\item{rec_power}{Recruitment power (for the power method)}

\item{rec_rate}{Recruitment rate (for the piecewise constant method)}

\item{rec_duration}{Recruitment duration (for the piecewise constant method)}
}
\value{
A dataframe
}
\description{
Function to simulate a data set, in the context of delayed treatment effects
}
