% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDTEDataSetPower.R
\name{SimDTEDataSetPower}
\alias{SimDTEDataSetPower}
\title{Function to simulate a data set, in the context of delayed treatment effects - using the power recruitment method}
\usage{
SimDTEDataSetPower(
  n1,
  n2,
  gammat,
  gammac,
  lambdat,
  lambdac,
  bigT,
  rec_period,
  rec_power
)
}
\arguments{
\item{n1}{Sample size in control group}

\item{n2}{Sample size in treatment group}

\item{gammat}{treatment parameter - gamma_t}

\item{gammac}{control parameter - gamma_c}

\item{lambdat}{treatment parameter - lambda_t}

\item{lambdac}{control parameter - lambda_c}

\item{bigT}{length of delay, T}

\item{rec_period}{recruitment period}

\item{rec_power}{recruitment power}
}
\value{
A dataframe
}
\description{
Function to simulate a data set, in the context of delayed treatment effects - using the power recruitment method
}
