% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDTEDataSetPWC.R
\name{SimDTEDataSetPWC}
\alias{SimDTEDataSetPWC}
\title{Function to simulate a data set, in the context of delayed treatment effects - using the piecewise constant recruitment method}
\usage{
SimDTEDataSetPWC(
  n1,
  n2,
  gammat,
  gammac,
  lambdat,
  lambdac,
  bigT,
  rec_rate,
  rec_duration
)
}
\arguments{
\item{n1}{Sample size in control group}

\item{n2}{Sample size in treatment group}

\item{gammat}{treatment parameter - gamma_t}

\item{gammac}{control parameter - gamma_c}

\item{lambdat}{treatment parameter - lambda_t}

\item{lambdac}{control parameter - lambda_c}

\item{bigT}{length of delay, T}

\item{rec_rate}{recruitment rate}

\item{rec_duration}{recruitment duration}
}
\value{
A dataframe
}
\description{
Function to simulate a data set, in the context of delayed treatment effects - using the piecewise constant recruitment method
}
