% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{survivalAnalysis}
\alias{survivalAnalysis}
\title{Function to test a data set for significance}
\usage{
survivalAnalysis(
  dataCombined,
  analysis_method = "LRT",
  alpha = 0.05,
  rho = 0,
  gamma = 0
)
}
\arguments{
\item{dataCombined}{A survival dataframe}

\item{analysis_method}{Method of analysis (log-rank test or weighted log-rank test)}

\item{alpha}{Type I error}

\item{rho}{Rho parameter for the Fleming-Harrington weighted log-rank test}

\item{gamma}{Gamma parameter for the Fleming-Harrington weighted log-rank test}
}
\value{
An indicator of significance
}
\description{
Function to test a data set for significance
}
