\name{ZODIAC}
\docType{data}
\alias{ZODIAC}
\title{ZODIAC clinical trial data set}
\description{
  Scraped Kaplan-Meier data from the clinical trial: Vandetanib plus docetaxel versus docetaxel as second-line treatment for patients with advanced non-small-cell lung cancer (ZODIAC): a double-blind, randomised, phase 3 trial
}
\usage{ZODIAC}
\format{A data set containing 697 observations.}
\references{
 Herbst RS, Sun Y, Eberhardt WE, et al. Vandetanib plus docetaxel versus docetaxel as second-line treatment for patients with
advanced non-small-cell lung cancer (ZODIAC): a double-blind, randomised, phase 3 trial. Lancet Oncol. 2010;11(7):619-
626. doi:10.1016/S1470-2045(10)70132-7
}
\keyword{datasets}
