% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_recruitment_time}
\alias{add_recruitment_time}
\title{Add a recruitment time to a survival data}
\usage{
add_recruitment_time(
  data,
  rec_method,
  rec_period = NULL,
  rec_power = NULL,
  rec_rate = NULL,
  rec_duration = NULL
)
}
\arguments{
\item{data}{A survival dataframe}

\item{rec_method}{Recruitment method, must be one of "power" or "PWC" (piecewise constant)}

\item{rec_period}{Parameter used to model recruitment according to power model}

\item{rec_power}{Parameter used to model recruitment according to power model}

\item{rec_rate}{Parameter used to model recruitment according to piecewise constant model}

\item{rec_duration}{Parameter used to model recruitment according to piecewise constant model}
}
\value{
a DTE data set
}
\description{
Add a recruitment time to a survival data
}
