% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_dte_assurance}
\alias{calc_dte_assurance}
\title{Function to simulate a data set, in the context of delayed treatment effects}
\usage{
calc_dte_assurance(
  n_c,
  n_t,
  lambda_c,
  HRStar,
  HRStarDist = "hist",
  gamma_C,
  gamma_E,
  delayT,
  delayTDist = "hist",
  P_S = 1,
  P_DTE = 0,
  censEvents = NULL,
  censTime = NULL,
  rec_method,
  rec_period = NULL,
  rec_power = NULL,
  rec_rate = NULL,
  rec_duration = NULL,
  analysis_method,
  rho = 0,
  gamma = 0,
  nSims = 10000
)
}
\arguments{
\item{HRStar}{The post-delay hazard ratio}

\item{HRStarDist}{The chosen parametric distribution for the post-delay hazard ratio}

\item{gamma_C}{Weibull control parameter}

\item{gamma_E}{Weibull experimental treatment parameter}

\item{delayT}{Length of delay}

\item{delayTDist}{The chosen parametric distribution for the length of delay}

\item{P_S}{Probability of the Kaplan-Meier curves separating at some time}

\item{P_DTE}{Probability of the treatment being subject to a delay (given the K-M curves will separate)}

\item{censEvents}{Number of events to be censored at}

\item{censTime}{Time of censoring}

\item{rec_method}{Recruitment method}

\item{rec_period}{Recruitment period (for the power method)}

\item{rec_power}{Recruitment power (for the power method)}

\item{rec_rate}{Recruitment rate (for the piecewise constant method)}

\item{rec_duration}{Recruitment duration (for the piecewise constant method)}

\item{analysis_method}{Method of analysis (log-rank test or weighted log-rank test)}

\item{rho}{Rho parameter for the Fleming-Harrington weighted log-rank test}

\item{gamma}{Gamma parameter for the Fleming-Harrington weighted log-rank test}

\item{nSims}{Number of simulations}

\item{n_C}{Sample size in control group}

\item{n_E}{Sample size in experimental treatment group}

\item{lambda_C}{Weibull control parameter}
}
\value{
A value
}
\description{
Function to simulate a data set, in the context of delayed treatment effects
}
