% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_dte_assurance}
\alias{calc_dte_assurance}
\title{Calculate Assurance for a trial with a Delayed Treatment Effect}
\usage{
calc_dte_assurance(
  n_c,
  n_t,
  lambda_c,
  gamma_c = NULL,
  control_dist = "Exponential",
  delay_time_SHELF,
  delay_time_dist = "hist",
  post_delay_HR_SHELF,
  post_delay_HR_dist = "hist",
  P_S = 1,
  P_DTE = 0,
  cens_method = "Time",
  cens_events = NULL,
  cens_time = NULL,
  rec_method,
  rec_period = NULL,
  rec_power = NULL,
  rec_rate = NULL,
  rec_duration = NULL,
  analysis_method = "LRT",
  alpha = 0.05,
  alternative = "one.sided",
  rho = 0,
  gamma = 0,
  nSims = 1000
)
}
\arguments{
\item{n_c}{Number of patients in the control group}

\item{n_t}{Number of patients in the treatment group}

\item{lambda_c}{Control group parameter}

\item{gamma_c}{Control group parameter}

\item{control_dist}{Distribution of control group, must be one of "Exponential" (default) or "weibull"}

\item{delay_time_SHELF}{A SHELF object, beliefs about the delay time}

\item{delay_time_dist}{Distribution of the delay time, "hist" is default. See SHELF help for more details}

\item{post_delay_HR_SHELF}{A SHELF object, beliefs about the post-delay hazard ratio}

\item{post_delay_HR_dist}{Distribution of the post-delay hazard ratio, "hist" is default. See SHELF help for more details}

\item{P_S}{Probability of the survival curves separating}

\item{P_DTE}{Probability of the survival curves being subject to a DTE, given they separate}

\item{cens_method}{Method of censoring, must be either "Time" (default) or "Events"}

\item{cens_events}{Number of events at which you wish to perform the censoring (must be less than n_c + n_t)}

\item{cens_time}{Time at which you wish to perform the censoring}

\item{rec_method}{Recruitment method, must be one of "power" or "PWC" (piecewise constant)}

\item{rec_period}{Parameter used to model recruitment according to power model}

\item{rec_power}{Parameter used to model recruitment according to power model}

\item{rec_rate}{Parameter used to model recruitment according to piecewise constant model}

\item{rec_duration}{Parameter used to model recruitment according to piecewise constant model}

\item{analysis_method}{Method of analysis, "LRT" (default) for standard log-rank test, or "WLRT" for weighted log-rank test}

\item{alpha}{Type I error}

\item{alternative}{String specifying the alternative hypothesis, must be one of "one.sided" or "two-sided" (default)}

\item{rho}{Rho parameter for the Fleming-Harrington weighted log-rank test}

\item{gamma}{Gamma parameter for the Fleming-Harrington weighted log-rank test}

\item{nSims}{Number of simulations, default is 1000}
}
\value{
A list: Assurance value, 95\% CI for assurance, Duration
}
\description{
Calculate Assurance for a trial with a Delayed Treatment Effect
}
