% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cens_data}
\alias{cens_data}
\title{Function to censor a data set}
\usage{
cens_data(data, cens_events = NULL, cens_time = NULL)
}
\arguments{
\item{data}{An uncensored dataframe}

\item{cens_events}{Number of events at which you wish to perform the censoring}

\item{cens_time}{Time at which you wish to perform the censoring}
}
\value{
A list: censored dataframe, cens_events and cens_time
}
\description{
Function to censor a data set
}
