% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cens_data}
\alias{cens_data}
\title{Censor a survival data set}
\usage{
cens_data(data, cens_method = "Time", cens_events = NULL, cens_time = NULL)
}
\arguments{
\item{data}{An uncensored dataframe}

\item{cens_method}{Method of censoring, must be either "Time" (default) or "Events"}

\item{cens_time}{Time at which you wish to perform the censoring}

\item{cens_IF}{Information Fraction at which you wish to perform the censoring}
}
\value{
A list: censored dataframe, cens_IF, cens_time and sample size
}
\description{
Censor a survival data set
}
