% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensFunc.R
\name{CensFunc}
\alias{CensFunc}
\title{Function to censor a data set}
\usage{
CensFunc(dataCombined, censEvents = NULL, censTime = NULL)
}
\arguments{
\item{dataCombined}{An uncensored dataframe}

\item{censEvents}{Number of events at which you wish to perform the censoring}

\item{censTime}{Time at which you wish to perform the censoring}
}
\value{
A censored dataframe
}
\description{
Function to censor a data set
}
