% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INTEREST}
\alias{INTEREST}
\title{INTEREST data set}
\format{
A data frame with 710 rows and 2 variables:
\describe{
\item{Survival time}{Survival Time (in months)}
\item{Status}{Event indicator (0=Alive, 1=Dead)}
}
}
\source{
Reconstructed survival data set from the following publication: https://www.sciencedirect.com/science/article/pii/S0140673608617584?via\%3Dihub
}
\usage{
INTEREST
}
\description{
A reconstructed survival data set for the INTEREST clinical trial
}
\keyword{datasets}
