% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REVEL}
\alias{REVEL}
\title{REVEL data set}
\format{
A data frame with 625 rows and 2 variables:
\describe{
\item{Survival time}{Survival Time (in months)}
\item{Status}{Event indicator (0=Alive, 1=Dead)}
}
}
\source{
Reconstructed survival data set from the following publication: https://www.thelancet.com/journals/lancet/article/PIIS0140-6736(14)60845-X/fulltext
}
\usage{
REVEL
}
\description{
A reconstructed survival data set for the REVEL clinical trial
}
\keyword{datasets}
