% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_dte}
\alias{sim_dte}
\title{Simulate a DTE data set}
\usage{
sim_dte(
  n_c,
  n_t,
  lambda_c,
  delay_time,
  post_delay_HR,
  dist = "exponential",
  gamma_c = NULL
)
}
\arguments{
\item{n_c}{number of patients in the control group}

\item{n_t}{number of patients in the treatment group}

\item{lambda_c}{control hazard}

\item{delay_time}{length of delay}

\item{post_delay_HR}{the hazard ratio, once the treatment begins to take effect}

\item{dist}{distribution of control distribution, must be one of "exponential" (default) or "weibull"}

\item{gamma_c}{}
}
\value{
a DTE data set
}
\description{
Simulate a DTE data set
}
