% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDTEDataSet.R
\name{SimDTEDataSet}
\alias{SimDTEDataSet}
\title{Simulates DTE data given certain inputs}
\usage{
SimDTEDataSet(
  n1,
  n2,
  gamma1,
  gamma2,
  lambda1,
  lambda2,
  bigT,
  recTime,
  censTime
)
}
\arguments{
\item{n1}{Number of patients in the control group.}

\item{n2}{Number of patients in the treatment group.}

\item{gamma1}{A parameter for the treatment.}

\item{gamma2}{A parameter for the control.}

\item{lambda1}{A parameter for the treatment.}

\item{lambda2}{A parameter for the control.}

\item{bigT}{The length of delay.}

\item{recTime}{The recruitment time.}

\item{censTime}{The time at which observations should be censored.}
}
\value{
A data frame consisting of the DTE data.
}
\description{
Simulates DTE data given certain inputs
}
\examples{
SimDTEDataSet(n1 = 100, n2 = 100, gamma1 = 0.8, gamma2 = 1.2,
lambda1 = 0.04, lambda2 = 0.08, bigT = 6, recTime = 6, censTime = 60)
}
