% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{survival_test}
\alias{survival_test}
\title{Test a data set for statistical significance}
\usage{
survival_test(
  data,
  analysis_method = "LRT",
  alternative = "one.sided",
  alpha = 0.05,
  rho = 0,
  gamma = 0
)
}
\arguments{
\item{data}{A survival dataframe}

\item{analysis_method}{Method of analysis, "LRT" (default) for standard log-rank test, or "WLRT" for weighted log-rank test}

\item{alternative}{String specifying the alternative hypothesis, must be one of "one.sided" or "two-sided" (default)}

\item{alpha}{Type I error}

\item{rho}{Rho parameter for the Fleming-Harrington weighted log-rank test}

\item{gamma}{Gamma parameter for the Fleming-Harrington weighted log-rank test}
}
\value{
An indicator of significance
}
\description{
Test a data set for statistical significance
}
