% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZODIAC}
\alias{ZODIAC}
\title{ZODIAC data set}
\format{
A data frame with 697 rows and 2 variables:
\describe{
\item{Survival time}{Survival Time (in months)}
\item{Status}{Event indicator (0=Alive, 1=Dead)}
}
}
\source{
Reconstructed survival data set from the following publication: https://www.sciencedirect.com/science/article/abs/pii/S1470204510701327?via\%3Dihub
}
\usage{
ZODIAC
}
\description{
A reconstructed survival data set for the ZODIAC clinical trial
}
\keyword{datasets}
